appindexer: Port the fuzzy scoring function to C
When transpiling to C, Vala really mangles this function, causing it to crash randomly on non-Latin text. Porting it to C and checking it against the reference implementation found online fixes the crashing issue while (as far as I can tell with Spanish) still producing the desired search results.

Another consideration was converting the search text and compare text to ASCII. This method also seems to fix the issue, though it feels more dirty. I decided to go forward with the C port because I simply can't trust Vala to do the right thing when it comes to this function.

Ref #383

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>